NDRS-942: Split account/delegator structs.
